% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_group_by.R
\name{count_group_by}
\alias{count_group_by}
\title{Count Group By}
\usage{
count_group_by(dataset, group_by_column, choose_value)
}
\arguments{
\item{dataset}{the data set that we choose to use}

\item{group_by_column}{the column that we choose to apply group_by function on}

\item{choose_value}{the value that we want to count}
}
\value{
A 1 x 2 tibble with value "choose_value" in column "group_by_column" and the count number n of this value
}
\description{
\code{count_group_by} returns the number of the value choose_value in the column group_by_column of the data set dataset
}
\examples{
library(dplyr)
library(datateachr)
library(tidyr)
count_group_by(building_permits, "issue_date", as.Date("2017-02-02", format = "\%Y-\%m-\%d"))
count_group_by(building_permits, "type_of_work", "New Building")
}
